requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-engineio==4.9.1
eventlet==0.36.0
gunicorn==22.0.0
gevent==24.11.1

python-dotenv==1.0.1

requests==2.32.3
qrcode==7.4.2
Pillow==10.4.0

setuptools==70.0.0